myproject/requirements.txt
amqp==5.2.0
asgiref==3.7.2
billiard==4.2.0
celery==5.3.6
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
dataclasses==0.6
Django
django-celery-results==2.5.1
django-ratelimit==4.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
djongo==1.2.31
dnspython==2.4.2
idna==3.6
install==1.3.5
kombu==5.3.5
prompt-toolkit==3.0.43
PyJWT==2.8.0
pymongo==4.6.1
python-dateutil==2.8.2
pytz==2023.3.post1
redis
redis-cli
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.1.0
vine==5.1.0
wcwidth==0.2.13